requirements.txt
crewai
fastapi
uvicorn
pydantic
python-dotenv
yfinance
pandas
langchain-tavily
crewai[azure-ai-inference]